deepfake/model/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow-cpu==2.19.0
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.26.4
Werkzeug==2.3.7
